backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
django-cors-headers>=4.3
chromadb>=0.5
pypdf>=4.0
google-genai>=1.0
python-dotenv>=1.0
sentence-transformers>=2.2.0
